Backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
PyMySQL==1.1.1
django-cors-headers==4.6.0
Pillow==11.0.0
python-decouple==3.8
requests==2.32.3
bcrypt==4.2.1
django-filter==24.3
openpyxl==3.1.5